Master Branch
There are 3 selection on the landing pages, Persistent, Animation and Music. On the persistent pages the user is allowed to purchase one of 3 type of products mild, medium and hot. After calculating the cost, the user is sent to another page to showcase the product image the purchase amount and the total cost. If the user slectes animation, the user can start a slideshow on another page. Though the user slectes the music option, the user can stop or start a music file.